Support both YOLOv5 (initial, no nms) and AutoML
